jsx => tsx and usage of memo
