+++ src/components/results/dynamicsimulation/dynamic-simulation-result-time-series.jsx
@@ -10,7 +10,7 @@ import CloseIcon from '@mui/icons-material/Close';
 import AddIcon from '@mui/icons-material/Add';
 import { Box, LinearProgress, Typography } from '@mui/material';
 import DynamicSimulationResultChart from './timeseries/dynamic-simulation-result-chart';
-import React, { useMemo, useState } from 'react';
+import React, { memo, useMemo, useState } from 'react';
 import { useIntl } from 'react-intl';
 import DroppableTabs from './common/draggable-tab/droppable-tabs';
 import DraggableTab from './common/draggable-tab/draggable-tab';
@@ -44,7 +44,7 @@ const styles = {
     },
 };
 
-const DynamicSimulationResultTimeSeries = ({ nodeUuid, studyUuid }) => {
+const DynamicSimulationResultTimeSeries = memo(({ nodeUuid, studyUuid }) => {
     const [result, loadTimeSeries, isLoading] = useResultTimeSeries(
         nodeUuid,
         studyUuid
@@ -216,7 +216,7 @@ const DynamicSimulationResultTimeSeries = ({ nodeUuid, studyUuid }) => {
             )}
         </>
     );
-};
+});
 
 DynamicSimulationResultTimeSeries.propTypes = {
     nodeUuid: PropTypes.string,
