+++ src/components/results/dynamicsimulation/dynamic-simulation-result-timeline.tsx
@@ -7,7 +7,7 @@
 import { UUID } from 'crypto';
 import { Box, LinearProgress } from '@mui/material';
 import { CustomAGGrid } from '../../custom-aggrid/custom-aggrid';
-import React, { useMemo, useRef } from 'react';
+import React, { memo, useMemo, useRef } from 'react';
 import { useIntl } from 'react-intl';
 import { makeAgGridCustomHeaderColumn } from '../../custom-aggrid/custom-aggrid-header-utils';
 import {
@@ -64,137 +64,141 @@ type DynamicSimulationResultTimelineProps = {
     nodeUuid: UUID;
 };
 
-const DynamicSimulationResultTimeline = ({
-    studyUuid,
-    nodeUuid,
-}: DynamicSimulationResultTimelineProps) => {
-    const intl = useIntl();
-    const gridRef = useRef(null);
+const DynamicSimulationResultTimeline = memo(
+    ({ studyUuid, nodeUuid }: DynamicSimulationResultTimelineProps) => {
+        const intl = useIntl();
+        const gridRef = useRef(null);
 
-    const [timelines, isLoading] = useNodeData(
-        studyUuid,
-        nodeUuid,
-        fetchDynamicSimulationResultTimeline,
-        dynamicSimulationResultInvalidations
-    );
+        const [timelines, isLoading] = useNodeData(
+            studyUuid,
+            nodeUuid,
+            fetchDynamicSimulationResultTimeline,
+            dynamicSimulationResultInvalidations
+        );
 
-    const { onSortChanged, sortConfig } = useAgGridLocalSort(gridRef, {
-        colKey: COL_TIME,
-        sortWay: SORT_WAYS.asc,
-    });
+        const { onSortChanged, sortConfig } = useAgGridLocalSort(gridRef, {
+            colKey: COL_TIME,
+            sortWay: SORT_WAYS.asc,
+        });
 
-    const { updateFilter, filterSelector } = useAggridLocalRowFilter(gridRef, {
-        [COL_TIME]: COL_TIME,
-        [COL_MODEL_NAME]: COL_MODEL_NAME,
-        [COL_MESSAGE]: COL_MESSAGE,
-    });
+        const { updateFilter, filterSelector } = useAggridLocalRowFilter(
+            gridRef,
+            {
+                [COL_TIME]: COL_TIME,
+                [COL_MODEL_NAME]: COL_MODEL_NAME,
+                [COL_MESSAGE]: COL_MESSAGE,
+            }
+        );
 
-    const sortAndFilterProps = useMemo(
-        () => ({
-            sortProps: {
-                onSortChanged,
-                sortConfig,
-            },
-            filterProps: {
-                updateFilter,
-                filterSelector,
-            },
-        }),
-        [onSortChanged, sortConfig, updateFilter, filterSelector]
-    );
-
-    // columns are defined from fields in {@link TimelineEvent} types
-    const columnDefs = useMemo(
-        () => [
-            makeAgGridCustomHeaderColumn({
-                headerName: intl.formatMessage({
-                    id: 'DynamicSimulationTimelineEventTime',
-                }),
-                field: COL_TIME,
-                width: MIN_COLUMN_WIDTH,
-                numeric: true,
-                fractionDigits: 2,
-                filter: 'agNumberColumnFilter',
-                filterParams: {
-                    filterDataType: FILTER_DATA_TYPES.NUMBER,
-                    filterComparators: Object.values(FILTER_NUMBER_COMPARATORS),
+        const sortAndFilterProps = useMemo(
+            () => ({
+                sortProps: {
+                    onSortChanged,
+                    sortConfig,
                 },
-                cellRenderer: NumberCellRenderer,
-                ...sortAndFilterProps,
-            }),
-            makeAgGridCustomHeaderColumn({
-                headerName: intl.formatMessage({
-                    id: 'DynamicSimulationTimelineEventModelName',
-                }),
-                field: COL_MODEL_NAME,
-                width: MEDIUM_COLUMN_WIDTH,
-                filterParams: {
-                    filterDataType: FILTER_DATA_TYPES.TEXT,
-                    filterComparators: [
-                        FILTER_TEXT_COMPARATORS.STARTS_WITH,
-                        FILTER_TEXT_COMPARATORS.CONTAINS,
-                    ],
+                filterProps: {
+                    updateFilter,
+                    filterSelector,
                 },
-                ...sortAndFilterProps,
             }),
-            makeAgGridCustomHeaderColumn({
-                headerName: intl.formatMessage({
-                    id: 'DynamicSimulationTimelineEventModelMessage',
+            [onSortChanged, sortConfig, updateFilter, filterSelector]
+        );
+
+        // columns are defined from fields in {@link TimelineEvent} types
+        const columnDefs = useMemo(
+            () => [
+                makeAgGridCustomHeaderColumn({
+                    headerName: intl.formatMessage({
+                        id: 'DynamicSimulationTimelineEventTime',
+                    }),
+                    field: COL_TIME,
+                    width: MIN_COLUMN_WIDTH,
+                    numeric: true,
+                    fractionDigits: 2,
+                    filter: 'agNumberColumnFilter',
+                    filterParams: {
+                        filterDataType: FILTER_DATA_TYPES.NUMBER,
+                        filterComparators: Object.values(
+                            FILTER_NUMBER_COMPARATORS
+                        ),
+                    },
+                    cellRenderer: NumberCellRenderer,
+                    ...sortAndFilterProps,
                 }),
-                field: COL_MESSAGE,
-                width: LARGE_COLUMN_WIDTH,
-                filterParams: {
-                    filterDataType: FILTER_DATA_TYPES.TEXT,
-                    filterComparators: [
-                        FILTER_TEXT_COMPARATORS.STARTS_WITH,
-                        FILTER_TEXT_COMPARATORS.CONTAINS,
-                    ],
-                },
-                ...sortAndFilterProps,
-            }),
-        ],
-        [intl, sortAndFilterProps]
-    );
+                makeAgGridCustomHeaderColumn({
+                    headerName: intl.formatMessage({
+                        id: 'DynamicSimulationTimelineEventModelName',
+                    }),
+                    field: COL_MODEL_NAME,
+                    width: MEDIUM_COLUMN_WIDTH,
+                    filterParams: {
+                        filterDataType: FILTER_DATA_TYPES.TEXT,
+                        filterComparators: [
+                            FILTER_TEXT_COMPARATORS.STARTS_WITH,
+                            FILTER_TEXT_COMPARATORS.CONTAINS,
+                        ],
+                    },
+                    ...sortAndFilterProps,
+                }),
+                makeAgGridCustomHeaderColumn({
+                    headerName: intl.formatMessage({
+                        id: 'DynamicSimulationTimelineEventModelMessage',
+                    }),
+                    field: COL_MESSAGE,
+                    width: LARGE_COLUMN_WIDTH,
+                    filterParams: {
+                        filterDataType: FILTER_DATA_TYPES.TEXT,
+                        filterComparators: [
+                            FILTER_TEXT_COMPARATORS.STARTS_WITH,
+                            FILTER_TEXT_COMPARATORS.CONTAINS,
+                        ],
+                    },
+                    ...sortAndFilterProps,
+                }),
+            ],
+            [intl, sortAndFilterProps]
+        );
 
-    // messages to show when no data
-    const dynamicSimulationStatus = useSelector(
-        (state: ReduxState) =>
-            state.computingStatus[ComputingType.DYNAMIC_SIMULATION]
-    );
-    const messages = useIntlResultStatusMessages(intl, true);
-    const overlayMessage = useMemo(
-        () =>
-            getNoRowsMessage(
-                messages,
-                timelines,
-                dynamicSimulationStatus,
-                !isLoading
-            ),
-        [messages, timelines, dynamicSimulationStatus, isLoading]
-    );
+        // messages to show when no data
+        const dynamicSimulationStatus = useSelector(
+            (state: ReduxState) =>
+                state.computingStatus[ComputingType.DYNAMIC_SIMULATION]
+        );
+        const messages = useIntlResultStatusMessages(intl, true);
+        const overlayMessage = useMemo(
+            () =>
+                getNoRowsMessage(
+                    messages,
+                    timelines,
+                    dynamicSimulationStatus,
+                    !isLoading
+                ),
+            [messages, timelines, dynamicSimulationStatus, isLoading]
+        );
 
-    const rowDataToShow = useMemo(
-        () => (overlayMessage ? [] : timelines),
-        [timelines, overlayMessage]
-    );
+        const rowDataToShow = useMemo(
+            () => (overlayMessage ? [] : timelines),
+            [timelines, overlayMessage]
+        );
 
-    return (
-        <>
-            {isLoading && (
-                <Box sx={styles.loader}>
-                    <LinearProgress />
-                </Box>
-            )}
-            <CustomAGGrid
-                ref={gridRef}
-                rowData={rowDataToShow}
-                columnDefs={columnDefs}
-                defaultColDef={defaultColDef}
-                overlayNoRowsTemplate={overlayMessage}
-                enableCellTextSelection
-            />
-        </>
-    );
-};
+        return (
+            <>
+                {isLoading && (
+                    <Box sx={styles.loader}>
+                        <LinearProgress />
+                    </Box>
+                )}
+                <CustomAGGrid
+                    ref={gridRef}
+                    rowData={rowDataToShow}
+                    columnDefs={columnDefs}
+                    defaultColDef={defaultColDef}
+                    overlayNoRowsTemplate={overlayMessage}
+                    enableCellTextSelection
+                />
+            </>
+        );
+    }
+);
 
 export default DynamicSimulationResultTimeline;
